feat(admin): enhance admin functionality with hotel and booking management features feat(booking): update booking controller to support hotel-specific bookings and improve validation feat(hotel): implement hotel management endpoints and integrate with user authentication feat(review): refine review controller to streamline review retrieval and management feat(room): add room management capabilities including image uploads and hotel association feat(user): enhance user controller with profile image upload and user retrieval
