--- src/booking/booking.controller.spec.ts
@@ -1,20 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { BookingController } from './booking.controller';
-import { BookingService } from './booking.service';
-
-describe('BookingController', () => {
-  let controller: BookingController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [BookingController],
-      providers: [BookingService],
-    }).compile();
-
-    controller = module.get<BookingController>(BookingController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
