--- src/review/review.controller.spec.ts
@@ -1,20 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { ReviewController } from './review.controller';
-import { ReviewService } from './review.service';
-
-describe('ReviewController', () => {
-  let controller: ReviewController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [ReviewController],
-      providers: [ReviewService],
-    }).compile();
-
-    controller = module.get<ReviewController>(ReviewController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
