--- src/room/room.controller.spec.ts
@@ -1,20 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { RoomController } from './room.controller';
-import { RoomService } from './room.service';
-
-describe('RoomController', () => {
-  let controller: RoomController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [RoomController],
-      providers: [RoomService],
-    }).compile();
-
-    controller = module.get<RoomController>(RoomController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
